AggregateRoot renamed to AggregateRootActor
