+++ SK_CHANGE_README.md
@@ -1,0 +1,2 @@
+Even if it will appear that one AR needs some resources from another (so in actors language it will mean calling actor)
+it can be (or even should be) expressed in some DDD service trait (interface).
