+++ src/test/scala/io/scalac/seed/route/VehicleRouteSpec.scala
@@ -2,7 +2,7 @@ package io.scalac.seed.route
 
 import akka.pattern.ask
 import akka.util.Timeout
-import io.scalac.seed.domain.AggregateRoot.Removed
+import io.scalac.seed.domain.AggregateRootActor.Removed
 import io.scalac.seed.domain.VehicleAggregate
 import io.scalac.seed.service.{UserAggregateManager, VehicleAggregateManager}
 import VehicleAggregate.Vehicle
